Move /routes into /server
